move all itests under all-itests module
