+++ hawkular-wildfly-agent-itest-parent/hawkular-wildfly-agent-all-itests/hawkular-wildfly-agent-cmdgw-itest/src/test/resources/org/hawkular/agent/ws/test/DatasourceCommandITest.testXaDs-after-add.node.txt
