+++ hawkular-wildfly-agent-itest-parent/hawkular-wildfly-agent-all-itests/hawkular-wildfly-agent-installer-itest/server-provisioning-hawkular.xml
