+++ hawkular-wildfly-agent-itest-parent/hawkular-wildfly-agent-all-itests/pom.xml
@@ -1,0 +1,64 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+
+    Copyright 2015-2016 Red Hat, Inc. and/or its affiliates
+    and other contributors as indicated by the @author tags.
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+
+-->
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+
+  <parent>
+    <groupId>org.hawkular.agent</groupId>
+    <artifactId>hawkular-wildfly-agent-itest-parent</artifactId>
+    <version>0.17.6.Final-SNAPSHOT</version>
+  </parent>
+
+  <artifactId>hawkular-wildfly-agent-all-itests</artifactId>
+  <packaging>pom</packaging>
+
+  <name>Hawkular Agent: Parent of All Integration Tests</name>
+
+  <properties>
+    <debug.failsafe.argLine />
+    <debug.wildfly.argLine />
+    <debug.host-controller.argLine />
+  </properties>
+
+  <modules>
+    <module>hawkular-wildfly-agent-installer-itest</module>
+    <module>hawkular-wildfly-agent-domain-itest</module>
+    <module>hawkular-wildfly-agent-cmdgw-itest</module>
+  </modules>
+
+  <profiles>
+    <profile>
+      <id>itest.debug</id>
+      <activation>
+        <property>
+          <name>itest.debug</name>
+        </property>
+      </activation>
+      <properties>
+        <debug.failsafe.argLine>-Xrunjdwp:transport=dt_socket,address=5007,server=y,suspend=y</debug.failsafe.argLine>
+        <debug.hawkular.argLine>-Xrunjdwp:transport=dt_socket,address=8787,server=y,suspend=n</debug.hawkular.argLine>
+        <!-- cannot get this to work - host controller gets errors
+        <debug.host-controller.argLine>-Xrunjdwp:transport=dt_socket,address=18787,server=y,suspend=n</debug.host-controller.argLine>
+        -->
+      </properties>
+    </profile>
+  </profiles>
+
+</project>
