+++ hawkular-wildfly-agent-itest-parent/hawkular-wildfly-agent-example-jndi/pom.xml
@@ -28,6 +28,7 @@
 
   <artifactId>hawkular-wildfly-agent-example-jndi</artifactId>
   <packaging>war</packaging>
+
   <name>Hawkular Agent: Example Custom Agent</name>
 
   <dependencies>
@@ -77,7 +78,7 @@
       <id>dev</id>
       <properties>
         <org.hawkular.dist.version>1.0.0.Alpha8</org.hawkular.dist.version>
-        <org.hawkular.wildfly.home>${basedir}/../../hawkular/dist/target/hawkular-${org.hawkular.dist.version}/hawkular-${org.hawkular.dist.version}</org.hawkular.wildfly.home>
+        <org.hawkular.wildfly.home>${basedir}/../../../hawkular/dist/target/hawkular-${org.hawkular.dist.version}/hawkular-${org.hawkular.dist.version}</org.hawkular.wildfly.home>
       </properties>
       <build>
         <plugins>
@@ -118,7 +119,7 @@
                   <mainClass>org.hawkular.wildfly.agent.installer.AgentInstaller</mainClass>
                   <arguments>
                     <argument>--target-location=${org.hawkular.wildfly.home}</argument>
-                    <argument>--module=${basedir}/../../hawkular-wildfly-agent-wf-extension/target/hawkular-wildfly-agent-wf-extension-${project.version}.zip</argument>
+                    <argument>--module=${basedir}/../../../hawkular-wildfly-agent-wf-extension/target/hawkular-wildfly-agent-wf-extension-${project.version}.zip</argument>
                     <argument>--server-url=http://localhost:8080</argument>
                     <argument>--username=jdoe</argument>
                     <argument>--password=password</argument>
